Refactor: restructure components in order to implement useParams hook.
