+++ src/App/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import { useState, useEffect } from 'react';
-import { fetchMovies, fetchSingleMovie, fetchSingleMovieVids } from '../apiCalls.js';
+import { fetchMovies } from '../apiCalls.js';
 import Movies from '../Movies/Movies.js';
 import MovieDetails from '../MovieDetails/MovieDetails.js';
 import Home from '../Home/Home'
@@ -10,8 +10,6 @@ import { Routes, Route, NavLink } from 'react-router-dom'
 function App() {
   const [movies, setMovies] = useState([]);
   const [error, setError] = useState('');
-  const [selectedMovie, setSelectedMovie] = useState(null);
-  const [selectedMovieVids, setSelectedMovieVids] = useState(null);
 
   useEffect(() => {
     const getMovies = async () => {
@@ -25,28 +23,6 @@ function App() {
     getMovies();
   }, [])
 
-  function selectMovie(id) {
-    setSelectedMovieVids('null')
-    const getSingleMovie = async () => {
-      try {
-        const details = await fetchSingleMovie(id);
-        setSelectedMovie( details.movie );
-      } catch (error) {
-        setError(error)
-      }
-    }
-    const getSingleMovieVids = async () => {
-      try {
-        const trailers = await fetchSingleMovieVids(id);
-        setSelectedMovieVids( trailers.videos );
-      } catch (error) {
-        setError(error)
-      }
-    }
-    getSingleMovieVids();    
-    getSingleMovie();
-  }
-
   return (
     <main className="App">
       <header>
@@ -57,21 +33,10 @@ function App() {
         </nav>
       </header>
       <Routes>
-        <Route path='/home' element={ <Home setSelectedMovie={setSelectedMovie} /> } />
-        <Route path='/movies' element={ movies.length ?
-          <Movies 
-            setSelectedMovie={setSelectedMovie} 
-            selectMovie={selectMovie} 
-            movies={movies} error={error} 
-          /> :
-          <NotFound error={error}/> } />
-        <Route path='/movies/:id' element={selectedMovie ? 
-          <MovieDetails 
-            selectedMovie={selectedMovie} 
-            selectedMovieVids={selectedMovieVids} 
-            error={error} 
-          /> : 
-          <NotFound error={error}/> } />
+        <Route path='/home' element={ <Home /> } />
+        <Route path='/movies' element={<Movies movies={movies} error={error} />} />
+        <Route path='/movies/:id' element={ <MovieDetails error={error} setError={setError}/> } />
+        <Route path='*' element={<NotFound />} />
       </Routes>
       <footer></footer>
     </main>
